+++ common/bazaar.h
@@ -1,0 +1,14 @@
+#ifndef EQEMU_BAZAAR_H
+#define EQEMU_BAZAAR_H
+
+#include <vector>
+#include "shareddb.h"
+
+class Bazaar {
+public:
+	static std::vector<BazaarSearchResultsFromDB_Struct>
+	GetSearchResults(SharedDatabase &db, BazaarSearchCriteria_Struct search, unsigned int char_zone_id);
+};
+
+
+#endif //EQEMU_BAZAAR_H
